+++ packages/terra-clinical-data-grid/src/subcomponents/HeaderCell.jsx
@@ -6,6 +6,7 @@ import IconCaretDown from 'terra-icon/lib/icon/IconCaretDown';
 
 import Cell from './Cell';
 import ResizeHandle from './ResizeHandle';
+import { sortIndicators } from '../proptypes/columnDataShape';
 
 import styles from './HeaderCell.module.scss';
 
@@ -29,7 +30,7 @@ const propTypes = {
    * rendered within in the header. However, if custom children are also provided, the sorting indicator will
    * not be rendered.
    */
-  sortIndicator: PropTypes.oneOf(['ascending', 'descending']),
+  sortIndicator: PropTypes.oneOf([sortIndicators.ASCENDING, sortIndicators.DESCENDING]),
   /**
    * Boolean indicating whether or not the HeaderCell is selectable.
    */
@@ -93,8 +94,8 @@ class HeaderCell extends React.Component {
       if (sortIndicator) {
         sortIndicatorComponent = (
           <div className={cx('sort-indicator')}>
-            {sortIndicator === 'ascending' ? <IconCaretUp /> : null}
-            {sortIndicator === 'descending' ? <IconCaretDown /> : null}
+            {sortIndicator === sortIndicators.ASCENDING ? <IconCaretUp /> : null}
+            {sortIndicator === sortIndicators.DESCENDING ? <IconCaretDown /> : null}
           </div>
         );
       }
@@ -129,5 +130,7 @@ class HeaderCell extends React.Component {
 }
 
 HeaderCell.propTypes = propTypes;
+HeaderCell.Opts = {};
+HeaderCell.Opts.SortIndicators = sortIndicators;
 
 export default HeaderCell;