+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added enumeration for valid column sort indicator values
 
 2.5.0 - (May 7, 2019)
 ----------